refactor(runner): remove dependency injection for inmemory repository
