+++ domain/runner/service.go
@@ -24,7 +24,7 @@ func (o Option) validate() error {
 	return nil
 }
 
-func NewService(p pipeline.IPipelineService, repo IRepositoryInmemory, o Option) (service IRunnerService, err error) {
+func NewService(p pipeline.IPipelineService, o Option) (service IRunnerService, err error) {
 	logger.Logger().Debug("-----------------------------------------------------------")
 	logger.Logger().Debug("START - Instantiate pipeline service")
 	logger.Logger().Debugf("< option = %+v", o)
@@ -45,7 +45,7 @@ func NewService(p pipeline.IPipelineService, repo IRepositoryInmemory, o Option)
 		return &runnerService{}, err
 	}
 
-	service = &runnerService{p, o, repo}
+	service = &runnerService{p, o, inmemoryRepository()}
 	return service, err
 }
 
@@ -53,5 +53,5 @@ type runnerService struct {
 	pipelineService pipeline.IPipelineService
 
 	option     Option
-	repository IRepositoryInmemory
+	repository iRepositoryInmemory
 }
